+++ layouts/GuidesLayout.tsx
@@ -2,9 +2,7 @@ import { ReactNode } from 'react'
 import { CoreContent } from 'pliny/utils/contentlayer'
 import ArticleLayout, { TocItemProps } from './ArticleLayout'
 import { RegionProvider } from '@/components/Region/RegionContext'
-import type { Guide } from '../types/transformedContent'
-
-type AuthorDetail = { name?: string; url?: string; image_url?: string; [key: string]: any }
+import type { AuthorDetail, Guide } from '../types/transformedContent'
 
 interface LayoutProps {
   content: CoreContent<Guide>
