+++ layouts/open-telemetry-hub/types.ts
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react'
 import type { CoreContent } from 'pliny/utils/contentlayer'
-import type { Blog, Comparison, Guide } from '../../types/transformedContent'
+import type { AuthorDetail, Blog, Comparison, Guide } from '../../types/transformedContent'
 import type { MDXContent } from '@/utils/strapi'
 
 export interface TocItemProps {
@@ -46,8 +46,6 @@ export type SidebarDoc = {
 }
 
 export type SidebarItem = SidebarCategory | SidebarDoc
-
-type AuthorDetail = { name?: string; url?: string; image_url?: string; [key: string]: any }
 
 export interface LayoutProps {
   content: CoreContent<Blog | Comparison | Guide | MDXContent>
